VEA-44 Refactored package structure.
Made error and exception handlers its own package, made filters their
own package.
