+++ src/main/java/com/danieldigiovanni/email/config/SecurityConfig.java
@@ -1,8 +1,8 @@
 package com.danieldigiovanni.email.config;
 
-import com.danieldigiovanni.email.config.filter.JwtAuthFilter;
-import com.danieldigiovanni.email.config.filter.LoggingFilter;
-import com.danieldigiovanni.email.config.filter.MetricsFilter;
+import com.danieldigiovanni.email.filter.JwtAuthFilter;
+import com.danieldigiovanni.email.filter.LoggingFilter;
+import com.danieldigiovanni.email.filter.MetricsFilter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
